refactor: 상태관리 시스템을 도메인별로 분리하여 폴더링 구조 개선
- cart, product, ui 도메인별로 독립적인 store 생성
- 각 도메인의 reducer와 초기 상태를 별도 파일로 분리
- productStore는 팩토리 패턴으로 초기 상태 주입 가능하도록 개선
